Inserting Address builder inside Address class
